Moving theme.ts from styling/src/utilities to styling/src/styles to prevent circular dependency.
