+++ packages/styling/src/styles/index.ts
@@ -3,3 +3,9 @@ export { DefaultPalette } from './DefaultPalette';
 export { DefaultFontStyles, FontSizes, FontWeights } from './DefaultFontStyles';
 export { IconCodes } from './IconCodes';
 export { getFocusStyle } from './getFocusStyle';
+export {
+  ITheme,
+  getTheme,
+  loadTheme,
+  createTheme
+} from './theme';
